this is a tmp commit (the real branch will be cleanup_bin) for testing xndarray unittests

+++ python/pyhrf/jde/jde_multi_sujets.py
@@ -2318,9 +2318,9 @@ class BOLDSampler_MultiSujInput:
 
     def buildParadigmConvolMatrix(self, zc, estimDuration, availableDataIndex,
                                   parData):
-        osf = self.tr / self.dt
+        osf = int(self.tr / self.dt)
 
-        logger.info('osf = %1.2f', osf)
+        logger.info('osf = %d', osf)
         logger.debug('availableDataIndex :')
         logger.debug(availableDataIndex)
